basic/log: use normal "insecure" getenv in log_parse_environment()
secure_getenv(3) says:

   The secure_getenv() function is intended for use in general-purpose
   libraries to avoid vulnerabilities that could occur if set-user-ID or
   set-group-ID programs accidentally trusted the environment.

log_parse_environment is only used internally in our daemons and is not
exported as a library function, or called by any of the library functions.
And it should not be: it uses systemd specific variable names.

So using secure_getenv() is not necessary. Our daemons are not written to be
used suid, and we should not pretend that anything like that is supported.

While at it, improve the error messages to mention what we were trying to
parse.

Follow-up for #4374.