ima: Write the policy filename into IMA's sysfs policy file
IMA validates file signatures based on the security.ima xattr. As of
Linux-4.7, instead of copying the IMA policy into the securityfs policy,
the IMA policy pathname can be written, allowing the IMA policy file
signature to be validated.

This patch modifies the existing code to first attempt to write the
pathname, but on failure falls back to copying the IMA policy contents.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>